refactor: 디자인패턴 변경으로인한 폴더구조 변경 및 로직 분리
