+++ src/components/container/coinDetail/DetailPrice.tsx
